% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{appendResults}
\alias{appendResults}
\title{Run results}
\usage{
appendResults(D, L, nmz = NULL, verbose = TRUE)
}
\arguments{
\item{D}{the PSA data}

\item{L}{a list of functions}

\item{nmz}{an optional vector specifying a subset of the functions in \code{F} to run}

\item{verbose}{print what's happening (default=\code{TRUE})}
}
\description{
Append a number of results to data
}
\author{
Pete Dodd
}
