% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getAB}
\alias{getAB}
\title{Get parameters for a beta distribution}
\usage{
getAB(E, V)
}
\arguments{
\item{E}{the mean}

\item{V}{the variance}
}
\value{
\code{list} containing a & b
}
\description{
Parameters for a beta distribution: a & b from
}
\author{
Pete Dodd
}
