% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getLNparms}
\alias{getLNparms}
\title{Get parameters for a log-normal distribution}
\usage{
getLNparms(mid, var, med = TRUE)
}
\arguments{
\item{mid}{the midpoint of the distribution (median/mean depending on \code{med})}

\item{var}{the variance of the distribution}

\item{med}{logical to assume \code{mid} is median (default) or otherwise mean.}
}
\value{
\code{list} mu, sig being mu and sdlog for lognormal distribution
}
\description{
Parameters for a lognormal distribution: mu & sdlog from median/mean and variance
}
\author{
Pete Dodd
}
