% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLabels.R
\name{GetNodeLabel}
\alias{GetNodeLabel}
\title{Labelling utilities}
\usage{
GetNodeLabel(node, varz = c("name", "cost", "qol"))
}
\arguments{
\item{node}{a \code{data.tree} object that must contain a probability field \code{p}}

\item{varz}{Variables}
}
\value{
\code{string} to be used as label
}
\description{
These are two labelling functions to label edges
and nodes used by \code{plotter}.
}
\author{
Pete Dodd
}
