% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hipnt}
\alias{hipnt}
\title{Utility function for extracting high point from bracketed uncertainty ranges}
\usage{
hipnt(x)
}
\arguments{
\item{x}{}
}
\value{
numeric
}
\description{
Expects something like M (Mlo,Mhi) or M (Mlo - Mhi)
}
\author{
Pete Dodd
}
