% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.xml.R
\name{insq}
\alias{insq}
\title{Utility to check if in a box}
\usage{
insq(p, X, Y, W, H, tol = 10)
}
\arguments{
\item{p}{2-vector location}

\item{X}{x value of top left}

\item{Y}{y value of top left (measured downwards)}

\item{W}{width}

\item{H}{height}

\item{tol}{a tolerance by which the box is expanded}
}
\value{
boolean
}
\description{
Utility to check if in a box
}
\author{
Pete Dodd
}
