% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.xml.R
\name{keyvalue}
\alias{keyvalue}
\title{Utility to split edge data on =}
\usage{
keyvalue(x)
}
\arguments{
\item{x}{string or vector thereof}
}
\value{
string or vector thereof
}
\description{
Utility to split edge data on =
}
\author{
Pete Dodd
}
