% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LabelFromData}
\alias{LabelFromData}
\title{Add/edit tree labels}
\usage{
LabelFromData(tree, data)
}
\arguments{
\item{tree}{a \code{data.tree} tree object}

\item{data}{a \code{data.frame} whose columns are the labels to be added}
}
\value{
This operates by side effect
}
\description{
This function will set (add or overwrite) tree label values based on
a data.frame whose columns are the labels to be added.
The rows of the data.frame should correspond to the way the tree is structured.
This can be used in conjunction with \code{tree2file} by adding columns to
CSV output, reading them in to be used as data input to this function.
}
\author{
Pete Dodd
}
