% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lopnt}
\alias{lopnt}
\title{Get low point from CIs}
\usage{
lopnt(x)
}
\arguments{
\item{x}{}
}
\value{
numeric
}
\description{
Utility function for extracting low point from bracketed uncertainty ranges
}
\details{
Expects something like M (Mlo,Mhi) or M (Mlo - Mhi)
}
\author{
Pete Dodd
}
