% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOdinmodel.R
\name{makeOdinModel}
\alias{makeOdinModel}
\title{Making Odin ODEs from tree}
\usage{
makeOdinModel(
  tree,
  rootname = "N1",
  rootinflow = "0",
  dimensions = c(),
  times.array = FALSE,
  quants = c()
)
}
\arguments{
\item{tree}{input tree}

\item{rootname}{base for naming convention}

\item{rootinflow}{if adding a flow in from existing model}

\item{dimensions}{the dimensions if an array model}

\item{times.array}{should the timescales also be arrays or not?}

\item{quants}{vector of entry-accruing quantities like cost to accumulate}
}
\value{
list
}
\description{
Making Odin ODEs from tree
}
\author{
Pete Dodd
}
