% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTfuns.R
\name{makeTfuns}
\alias{makeTfuns}
\title{Make functions to evaluate tree}
\usage{
makeTfuns(node, qnt = c("cost", "qol"))
}
\arguments{
\item{node}{a \code{data.tree} object that must contain a probability field \code{p}}

\item{qnt}{Quantity}
}
\value{
A \code{list} of functions
}
\description{
This function prepares functions that evaluate the strings generated by \code{getAQ}.
The default quantities for mean calculation over tree are `cost` and `qol`,
but others can be specified and will work as long as these
variables are defined at all nodes of the tree.
}
\author{
Pete Dodd
}
