% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{midpnt}
\alias{midpnt}
\title{Utility function for extracting midpoint from bracketed uncertainty ranges}
\usage{
midpnt(x)
}
\arguments{
\item{x}{String}
}
\value{
numeric
}
\description{
Expects something like M (Mlo,Mhi) or M (Mlo - Mhi)
}
\author{
Pete Dodd
}
