% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org2tree.R
\name{MSorg2tree}
\alias{MSorg2tree}
\title{Text file to \code{data.tree} tree}
\usage{
MSorg2tree(fn)
}
\arguments{
\item{fn}{the filename}
}
\value{
A \code{data.tree} tree object
}
\description{
This function reads an MS word organisation chart (saved as plain text)
file and generates a \code{data.tree} tree object.
This enables rapid specification of and experimentation with tree logics.
Different levels of the hierarchy are specified in the text file by the
number of tabs the line begins with.
}
\author{
Pete Dodd
}
