% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.xml.R
\name{nobr}
\alias{nobr}
\title{Utility nobr strips <br>}
\usage{
nobr(x)
}
\arguments{
\item{x}{string or vector thereof}
}
\value{
string or vector thereof
}
\description{
Utility nobr strips <br>
}
\author{
Pete Dodd
}
