% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.xml.R
\name{nospace}
\alias{nospace}
\title{Utility strips space etc}
\usage{
nospace(x)
}
\arguments{
\item{x}{string or vector thereof}
}
\value{
string or vector thereof
}
\description{
Utility strips space etc
}
\author{
Pete Dodd
}
