% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org2tree.R
\name{org2LoL}
\alias{org2LoL}
\title{Emacs org-mode file to list-of-lists}
\usage{
org2LoL(fn)
}
\arguments{
\item{fn}{the filename}
}
\value{
A \code{list} of \code{list}s
}
\description{
This function reads and Emacs org-mode file and generates a list-of-lists. This is used primarily as an intermediate: \code{org2tree} returns a \code{data.tree} directly.
}
\author{
Pete Dodd
}
