% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org2tree.R
\name{org2tree}
\alias{org2tree}
\title{Emacs org-mode file to \code{data.tree} tree}
\usage{
org2tree(fn)
}
\arguments{
\item{fn}{the filename}
}
\value{
A \code{data.tree} tree object
}
\description{
Reads and Emacs org-mode file and generates a \code{data.tree} tree object.
This enables rapid specification of and experimentation with tree logics.
Different levels of the hierarchy as specified by the number of asterisks
the line begins with (a space must follow the asterisks).
Emacs provides folding and other tools for rapidly editing org-mode files.
}
\author{
Pete Dodd
}
