% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.parmtable.R
\name{parse.parmtable}
\alias{parse.parmtable}
\title{Reading a parameter table and making utility functions}
\usage{
parse.parmtable(data, outfile = NULL, dpl = 3, testdir = NULL)
}
\arguments{
\item{data}{}

\item{outfile}{}

\item{dpl}{}

\item{testdir}{}
}
\value{
\code{list} of functions for RV computations and associated names
}
\description{
This function is designed to simplify the workflow by allowing specification of parameters/priors in an external CSV file following certain conventions.
}
\details{
Work in progress: need to add other distributions; testing functionality etc
}
\author{
Pete Dodd
}
