% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PruneByOutcome.R
\name{PruneByOutcome}
\alias{PruneByOutcome}
\title{Prune Trees by Outcome}
\usage{
PruneByOutcome(TREE, outcome, negate = FALSE)
}
\arguments{
\item{TREE}{an input tree}

\item{outcome}{a string for a variable that can be interpreted as >0 on positive outcome leaves}

\item{negate}{whether the indicator should be negated}
}
\value{
a pruned tree only containing routes to leaves with positive outcomes
}
\description{
Utility to subset trees by leaf outcome
}
\details{
The purpose of this utility is to allow calculation of tree averages conditional on a specified outcome. This has particular usefulness if a tree is to calculate transitions for use in with other models (e.g. ODEs or Markov models)
}
\author{
Pete Dodd
}
