% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.xml.R
\name{read.xml}
\alias{read.xml}
\title{Reads a tree and returns}
\usage{
read.xml(fn, AllowDuplicateNodes = FALSE)
}
\arguments{
\item{fn}{The filename of the file to be read.}
}
\value{
A data.tree of with specified node names and data.
}
\description{
This function is designed to read in trees from draw.io exported as
(uncompressed) xml
}
\details{
Node names are arbitrary (but <br> is stripped from strings).
Edge data is best specified as p=1,c=c2 etc. p is treated as a probability.
Here a variable called c is created, but a different name and an arbitrary number
of other variables can be added. Whitespace is stripped.
A utility variable called check is added to leaves to facilitate testing/debugging.
}
\author{
Pete Dodd
}
