% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{runningprobs}
\alias{runningprobs}
\title{Attach probabilities of reaching each node}
\usage{
runningprobs(tree)
}
\arguments{
\item{tree}{input tree}
}
\value{
by side-effect on tree as runningprob
}
\description{
Attach probabilities of reaching each node
}
\author{
Pete Dodd
}
