% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showParmz.R
\name{showParmz}
\alias{showParmz}
\title{List parameters in model}
\usage{
showParmz(node, parmz = c("cost", "p", "qol"))
}
\arguments{
\item{node}{a \code{data.tree} object that must contain a probability field \code{p}}

\item{parmz}{}
}
\value{
\code{list} of length 2: unique basic parameters & unique calculation-defined parameters
}
\description{
This function just fetches the parameters used in the model (and prints them).
It separates those which are basic and require separate definition from
those that are specified via calculations.
}
\author{
Pete Dodd
}
