% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate}
\alias{simulate}
\title{Simulate numbers on the tree}
\usage{
simulate(tree, N, P)
}
\arguments{
\item{tree}{input tree}

\item{N}{integer inflow}

\item{P}{list of data.frame/table row of parameters}
}
\value{
by side-effect on tree as N
}
\description{
Simulation of tree
}
\details{
This is for simulating numbers from studies. The tree can be converted afterwards with something like

ToDataFrameTable(RStbtxo, "pathString", "N")

which is part of the data.tree library
}
\author{
Pete Dodd
}
