% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tree2file}
\alias{tree2file}
\title{Write a CSV tree with labels}
\usage{
tree2file(TREE, filename, ...)
}
\arguments{
\item{TREE}{the tree}

\item{filename}{file to write to}

\item{...}{label names to include}
}
\description{
Write a CSV tree with labels
}
\author{
Pete Dodd
}
