% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{txt2tree}
\alias{txt2tree}
\title{Make a tree from a tsv}
\usage{
txt2tree(x)
}
\arguments{
\item{x}{filename relative to 'here'}
}
\value{
A tree
}
\description{
Simpler text file to tree
}
\details{
See also MSorg2tree
}
\author{
Pete Dodd
}
