% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lopnt}
\alias{lopnt}
\title{Utility function for extracting low point from bracketed uncertainty ranges}
\usage{
lopnt(x)
}
\arguments{
\item{x}{String}
}
\value{
numeric
}
\description{
Expects something like M (Mlo,Mhi) or M (Mlo - Mhi)
}
\author{
Pete Dodd
}
