% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOdinmodel.R
\name{makeOdinModel}
\alias{makeOdinModel}
\title{Making Odin ODEs from tree}
\usage{
makeOdinModel(tree, basename = "N1")
}
\arguments{
\item{tree}{}

\item{basename}{}
}
\value{
list
}
\description{
Making Odin ODEs from tree
}
\author{
Pete Dodd
}
