% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showParmz.R
\name{makeTestData}
\alias{makeTestData}
\title{List parameters in model}
\usage{
makeTestData(ncheck, vnames)
}
\arguments{
\item{node}{}
}
\value{
\code{data.frame} of dummy PSA for testing tree functions
}
\description{
This function makes a dummy PSA dataset based on a set of parameters for function testing
}
\author{
Pete Dodd
}
