% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{MergeByName}
\alias{MergeByName}
\title{Merge a tree onto another by node name}
\usage{
MergeByName(rootnode, nodetoadd, nodename, usecase = TRUE, leavesonly = FALSE)
}
\arguments{
\item{rootnode}{}

\item{nodetoadd}{}

\item{nodename}{}

\item{usecase}{Match case in nodename? (must be TRUE currently)}

\item{leavesonly}{Merge only onto leaves?}
}
\description{
Merge a tree onto another by node name
}
\author{
Pete Dodd
}
