% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{MergeByName}
\alias{MergeByName}
\title{Merge a tree onto another by node name}
\usage{
MergeByName(rootnode, nodetoadd, nodename, usecase = TRUE)
}
\arguments{
\item{rootnode}{}

\item{nodetoadd}{}

\item{nodename}{}

\item{usecase}{}
}
\description{
Merge a tree onto another by node name
}
\author{
Pete Dodd
}
