% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{midpnt}
\alias{midpnt}
\title{Get midpoint from CIs}
\usage{
midpnt(x)
}
\arguments{
\item{x}{}
}
\value{
numeric
}
\description{
Utility function for extracting midpoint from bracketed uncertainty ranges
}
\details{
Expects something like M (Mlo,Mhi) or M (Mlo - Mhi)
}
\author{
Pete Dodd
}
