% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.parmtable.R
\name{parse.parmtable}
\alias{parse.parmtable}
\title{Reading a parameter table and making utility functions}
\usage{
parse.parmtable(
  data,
  outfile = NULL,
  dpl = 3,
  testdir = NULL,
  qn = 0.75,
  tgphfmt = "png"
)
}
\arguments{
\item{data}{in data, first col parm name, second definition}

\item{outfile}{optional name for test table out}

\item{dpl}{number of dps for reporting (TODO)}

\item{testdir}{test directory. Settings this acts as a flag for action. TODO}

\item{qn}{quantiles to report in table TODO}

\item{tgphfmt}{test graph format (default 'png', otherwide 'pdf')}
}
\value{
\code{list} of functions for RV computations and associated names
}
\description{
This function is designed to simplify the workflow by allowing specification
of parameters/priors in an external CSV file following certain conventions.
}
\details{
Work in progress: need to add other distributions; testing functionality etc


The following distributions are handled:

\itemize{
 \item{"Normal"}{defined as \code{N(mean,sd)}}
 \item{"Log-Normal"}{defined as \code{LN(mu,sdlog)}}
 \item{"Beta"}{defined as \code{B(a,b)}}
 \item{"Exponential"}{defined as \code{E(rate)}}
}

TODO document & expand inputs below,
}
\examples{
tst <- parse.parmtable(data.frame(parm=c('blah','nla'),dist=c('LN(2,3)','N(2,3)')))
tst[[1]]$r(10)
tst[['blah']]$r(10)

}
\author{
Pete Dodd
}
