% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showParmz.R
\name{showAllParmz}
\alias{showAllParmz}
\title{List parameters in model}
\usage{
showAllParmz(TREE)
}
\arguments{
\item{node}{}
}
\value{
\code{vector} of  unique basic parameters
}
\description{
This function just fetches the parameters used in the model for use in testing.
}
\author{
Pete Dodd
}
