% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{txt2tree}
\alias{txt2tree}
\title{Simpler text file to tree}
\usage{
txt2tree(x)
}
\arguments{
\item{x}{Filename relative to 'here'}
}
\value{
A tree
}
\description{
See also MSorg2tree
}
\author{
Pete Dodd
}
